Do not Auto Build Game Packages in Debug Mode (#13)
The Packaging Nuget will auto build the various game platform packages by default. So in debug mode we need to make sure we don't do that.

This is done by setting the AutoBuildPackages property to False in the Debug configuration in the csproj. Only the Desktop csproj uses the Packaging Nuget so its the only project that needs to be modified.